pyproject.toml
[tool.poetry]
name = "jwt_producer"
version = "0.1.0"
description = ""
authors = ["likeinlife <likeinlife@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyJWT = {version = "2.8.0", extras = ["crypto"]}

[tool.poetry.dev-dependencies]
ruff = "^0.1.6"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
